Create component for ToDoList and refactor App.js
